Sec(6)-PyTorch/ops/gradioapp/requirements.txt
gradio[client]==4.44.1
torch 
torchvision
opencv-python
pillow
